header: add setting to escape the match template
Previously we always escaped the match template, unless
MatchTemplateEscapeDisable was set to `true`.

I think it makes more sense to do this the opposite way around, and
expect a regexp by default and allow the user to enable escaping. If a
match template is not provided, we always escape the template value.