cmd/golicenser: fix flags missing
The init in cmd/golicenser was removed in #2 to comply with
golangci-lint's rule of "It must not contain init()", however the
replacement implementation was broken due to the flags being created
after they would be parsed.

As the init is in the main package, I don't believe it will be an issue
for golangci-lint. This also fixes us using flag instead of flagSet.