header: rename MatchTemplate to Matcher
Matcher is used to match the project's license header, meaning any
headers that match should be updated or replaced by golicenser.

MatchTemplate made sense at the time, as it is a template which is
executed then used as a regexp matcher, however I believe Matcher makes
more sense as it is used to match the header.

Matcher is still a template which is executed, as such this should be
well documented.