analysis: disable concurrency in analyzer by default
golangci-lint handles concurrency internally, meaning we should avoid
creating an unnecessary goroutine and instead allow MaxConcurrent to be
set to 0, meaning files should be processed synchronously.

This also optimises for when MaxConcurrent is set to 1, since it is the
same as using no concurrency and we can skip using goroutines.

cmd/golicenser keeps the previous default of `runtime.NumCPU() * 2` for
the MaxConcurrent value.